move CapabilityRequest into core_types
